Update text.simba
slushpuppy had committed a line in ChooseOptionMultiEx (specifically the ClickLeft case action), in which broke the use-ability for ``child'' functions.  The code is reverted back to what works properly, and should probably stay as is next time unless there's a conflict with something specific on slushpuppy's/anyone else's end aside from saying it's fixed (in my opinion). :)