Add ChooseOption.SelectID, tweak how & when the waits are applied
A lot of code changed here, but it seemed to work exactly as expected, without any glitches.

Main change is that before, we did a wait right after opening the menu, now instead we create a "fake" estimate goal location, move the mouse there, do the wait now instead, and if that estimate is off we move the mouse to the correct option, then finally we click it. It's marginally more realistic.