Update time.simba
The if statement on line 171 was preventing the TTimeMarker from setting FPrevMark to the tick count in TTimeMarker.Start() when the timer was resumed from a paused state. This causes TTimeMarker.GetTime() to return an incorrect value after the TTimeMarker is started again because it uses FPrevMark to calculate the time since the Start() function was called.