Update IsOpen and Close in ChooseOption.simba
Changed "IsOpen" to use "Find" instead of "__Find" so that the option box boundaries will be set after "IsOpen" is called. I don't think this should cause issues since "Find" has a call to the other "__Find" function inside of it.

In "Close" function I simplified the expression on line 412 of original to the equivalent expression: (dir.T-Pi/3) + Random()*(2*Pi/3);
In "Close" function, I also added a piece of code to detect when the option box is within a few pixels of the left boundary of the client and over 500 pixels tall. If the option box is all the way to the left, you can not clear it by moving the mouse that direction, and if the option box is taller than 500 pixels it usually means the "Cancel" option has been bumped off the bottom, so the only way to clear it is to move the mouse to the right. This code snippet will force the move direction to an angle between the top right corner and bottom right corner of the option box, when the previous conditions are detected.